Move layouts directory into components.
