+++ myuw/static/vue/components/cards/husky-experience.vue
@@ -41,7 +41,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../layouts/card.vue';
+import Card from '../../components/layouts/card.vue';
 
 export default {
   components: {
