+++ myuw/static/vue/components/cards/outage.vue
@@ -69,7 +69,7 @@
 
 <script>
 import {mapGetters, mapState} from 'vuex';
-import Card from '../../layouts/card.vue';
+import Card from '../../components/layouts/card.vue';
 
 export default {
   components: {
