+++ myuw/static/vue/components/cards/schedule/visual-schedule.vue
@@ -44,7 +44,7 @@
 <script>
 import dayjs from 'dayjs';
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../layouts/card.vue';
+import Card from '../../../components/layouts/card.vue';
 import ScheduleTab from './schedule-tab.vue';
 
 export default {
