+++ myuw/static/vue/components/index/cards/accounts/uwnetid.vue
@@ -37,7 +37,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
