+++ myuw/static/vue/components/index/cards/applicant/bothell.vue
@@ -73,7 +73,7 @@
 </template>
 
 <script>
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
