+++ myuw/static/vue/components/index/cards/applicant/seattle.vue
@@ -222,7 +222,7 @@
 </template>
 
 <script>
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
