+++ myuw/static/vue/components/index/cards/applicant/tacoma.vue
@@ -178,7 +178,7 @@
 </template>
 
 <script>
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
