+++ myuw/static/vue/components/index/cards/events/events.vue
@@ -109,7 +109,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 import ListEvents from './list-events.vue';
 
 export default {
