+++ myuw/static/vue/components/index/cards/international/new-student.vue
@@ -21,7 +21,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
