+++ myuw/static/vue/components/index/cards/new_student/critical-info.vue
@@ -59,7 +59,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
