+++ myuw/static/vue/components/index/cards/new_student/to-register.vue
@@ -136,7 +136,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 
 export default {
   components: {
