+++ myuw/static/vue/components/index/cards/notices.vue
@@ -78,7 +78,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../layouts/card.vue';
+import Card from '../../../components/layouts/card.vue';
 
 export default {
   components: {
