+++ myuw/static/vue/components/index/cards/quicklinks/quicklinks.vue
@@ -127,7 +127,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 import Link from './link.vue';
 import CovidLinks from './covid-links.vue';
