+++ myuw/static/vue/components/index/cards/registration/status.vue
@@ -112,7 +112,7 @@
 import dayjs from 'dayjs';
 import {mapGetters, mapState, mapActions} from 'vuex';
 
-import Card from '../../../../layouts/card.vue';
+import Card from '../../../../components/layouts/card.vue';
 import EstRegComponent from './estRegDate.vue';
 import FinAidComponent from './finaid.vue';
 import HoldsComponent from './holds.vue';
